Refactor DynamoDB configuration and repository access in Instant Payments service
- Updated environment variables in `.env.example` and scripts to use `AWS_ENDPOINT_URL` and `AWS_REGION` instead of `DYNAMODB_ENDPOINT` and `DYNAMODB_REGION`.
- Refactored the DynamoDB repository access to support both high-level and low-level clients, allowing for more flexibility in data handling.
- Adjusted the application logic to utilize the new repository structure, enhancing maintainability and clarity.
- Removed deprecated files related to the old DynamoDB access methods to streamline the codebase.
