+++ examples/sample-apps/instant-payments/nodejs/.env.example
@@ -1,18 +1,19 @@
 PORT=8080
 
+# Required DynamoDB connectivity (use http://localhost:18000 when tests or CLI run on the host against compose below)
+# Standard AWS endpoint override:
+AWS_ENDPOINT_URL=http://localhost:18000
+# Standard AWS region:
+AWS_REGION=eu-west-1
+
 # Host port published for DynamoDB Local in docker-compose (container still uses 8000).
 # Use 18000 if something else (e.g. SSH) already binds host port 8000.
 DYNAMODB_LOCAL_HOST_PORT=18000
-
-# Required DynamoDB connectivity (use http://localhost:18000 when tests or CLI run on the host against compose below)
-DYNAMODB_ENDPOINT=http://localhost:18000
-DYNAMODB_REGION=eu-west-1
-DYNAMODB_CLIENTTYPE=high-level
-DYNAMODB_TABLENAME=JS_InstantPayments
-
+# DynamoDB client type: high-level or low-level
+DYNAMODB_CLIENT_TYPE=high-level
+DYNAMODB_TABLE_NAME=JS_InstantPayments
 # Required by spec: inclusive range 3600..31536000
 DYNAMODB_IDEMPOTENCY_TTL_SECONDS=2592000
-
 # Optional (LATEST default). Tests may use TRIM_HORIZON.
 DYNAMODB_STREAMS_ITERATOR_TYPE=LATEST
